requirements.txt
Pillow==2.6.1
certifi==14.05.14
colorama==0.3.2
facebook-sdk==1.0.0-alpha
pymongo==2.7.1
requests==2.4.3
tornado==4.0.2
